+++ CHANGELOG.md
@@ -15,3 +15,16 @@ All notable changes to the "vscode-modaledit" extension will be documented in th
 
 - Added `startInNormalMode` setting, resolving issue 
   [#1](https://github.com/johtela/vscode-modaledit/issues/1)
+
+
+## Version 1.3
+
+- Incremental search now returns to insert mode, if it was invoked from there 
+  [#4](https://github.com/johtela/vscode-modaledit/issues/4)
+- Added new command `modaledit.typeNormalKeys` which can be used to "call"
+  key bindings. Also fixes issue 
+  [#3](https://github.com/johtela/vscode-modaledit/issues/3)
+- Added new argument `typeAfterAccept` to `modaledit.search` command. This 
+  invokes normal mode key bindings (using `modaledit.typeNormalKeys`) after
+  successful search. The argument can be used to enter insert mode, or clear 
+  selection after search, for example.